pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "google-genai==1.12.1",
    "langchain>=1.2.11",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.2.0",
    "langchain-text-splitters>=1.1.1",
    "python-dotenv==1.1.0",
    "streamlit>=1.55.0",
]
